requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==3.2.5
django-environ==0.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
pyotp==2.7.0
